refact: split the mapPopup and mapTiff components
